+++ src/main/java/com/clubber/ClubberServer/domain/admin/service/AdminAuthService.java
@@ -1,0 +1,57 @@
+package com.clubber.ClubberServer.domain.admin.service;
+
+import com.clubber.ClubberServer.domain.admin.domain.Admin;
+import com.clubber.ClubberServer.domain.admin.dto.CreateAdminsLoginRequest;
+import com.clubber.ClubberServer.domain.admin.dto.CreateAdminsLoginResponse;
+import com.clubber.ClubberServer.domain.admin.validator.AdminValidator;
+import com.clubber.ClubberServer.domain.user.domain.RefreshTokenEntity;
+import com.clubber.ClubberServer.domain.user.exception.RefreshTokenExpiredException;
+import com.clubber.ClubberServer.domain.user.repository.RefreshTokenRepository;
+import com.clubber.ClubberServer.global.config.security.SecurityUtils;
+import com.clubber.ClubberServer.global.jwt.JwtTokenProvider;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class AdminAuthService {
+
+	private final AdminReadService adminReadService;
+	private final JwtTokenProvider jwtTokenProvider;
+	private final RefreshTokenRepository refreshTokenRepository;
+	private final AdminValidator adminValidator;
+
+	@Transactional
+	public CreateAdminsLoginResponse createAdminsLogin(CreateAdminsLoginRequest loginRequest) {
+		Admin admin = adminReadService.getAdminByUsername(loginRequest.getUsername());
+		adminValidator.validatePassword(loginRequest.getPassword(), admin.getPassword());
+		return createAdminsToken(admin);
+	}
+
+	private CreateAdminsLoginResponse createAdminsToken(Admin admin) {
+		String accessToken = jwtTokenProvider.generateAccessToken(admin);
+		String refreshToken = jwtTokenProvider.generateRefreshToken(admin.getId());
+		RefreshTokenEntity refreshTokenEntity = RefreshTokenEntity.of(admin.getId(), refreshToken,
+			jwtTokenProvider.getRefreshTokenTTlSecond());
+		RefreshTokenEntity savedRefreshToken = refreshTokenRepository.save(refreshTokenEntity);
+		return CreateAdminsLoginResponse.of(admin, accessToken,
+			savedRefreshToken.getRefreshToken());
+	}
+
+	@Transactional
+	public CreateAdminsLoginResponse getAdminsParseToken(String refreshToken) {
+		RefreshTokenEntity refreshTokenEntity = refreshTokenRepository.findByRefreshToken(
+				refreshToken)
+			.orElseThrow(() -> RefreshTokenExpiredException.EXCEPTION);
+		Long adminId = jwtTokenProvider.parseRefreshToken(refreshTokenEntity.getRefreshToken());
+		Admin admin = adminReadService.getAdminById(adminId);
+		return createAdminsToken(admin);
+	}
+
+	@Transactional
+	public void logout() {
+		Long currentUserId = SecurityUtils.getCurrentUserId();
+		refreshTokenRepository.deleteById(currentUserId);
+	}
+}
